2021-12-22 Specification build from map

+++ src/main/java/com/liaoxx/spring_hello/dto/api/goods/GoodsListDto.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class GoodsListDto implements DtoI {
-    public int count=0;
+    public long count=0;
 
     public int page=0;
